feat(pollutants): baseline scaffolding for Largo test scenario
Wires up the data and plumbing for pollutant dynamics so we can start
tuning visually. Numbers, weights, and the colour formula are all
placeholders — this is the baseline scaffolding for the test scenario,
not a tuned simulation.

- 8 entities (sewage/agri e.coli, phosphate, sediment, litter, seal,
  oyster, seagrass) with per-zone baselines, annual additions and
  declines, sourced from the CSVs in _temporary
- New RoundSchedule (Addition/Decline/Nothing) on Scenario, applied at
  the start of each round before movement runs
- Estuary regions alias to the nearest freshwater compute cell so they
  share state
- One-way pollutant movement: freshwater->seawater, overflow->seawater
- PollutantColorDriver tints water materials per-zone via a weighted
  norm (eColi half-weighted), lerped over 2s in WaterTintController.
  Freshwater muddies, seawater stays mostly blue (dilution)
- Population storage widened to long/double so e.coli values (~1e14)
  don't overflow
- HiddenFromCellToken flag on Entity hides sealife tokens on per-cell
  view while keeping them in the side panel

+++ Assets/Scripts/Sandbox/Maths/IEntityCalculator.cs
@@ -7,7 +7,7 @@ namespace Glitchers.EcoKnow.Sandbox
         public string Name();
         public string Version();
 
-        public void CalculatePopulations(EntityManager entityManager, int[,,] entityLookupTable);
-        public void CalculateMovement(EntityManager entityManager, int[,,] entityLookupTable);
+        public void CalculatePopulations(EntityManager entityManager, long[,,] entityLookupTable);
+        public void CalculateMovement(EntityManager entityManager, long[,,] entityLookupTable);
     }
 }
